requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
psutil==6.0.0
python-dotenv==1.0.1